// src/modules/authors/dto/authors.dto.ts
export default class AuthorDto {
  sintaId?: string = null;
  name?: string = null;
  profileUrl?: string = null;
  profileImgUrl?: string = null;
  universityName?: string = null;
  universityUrl?: string = null;
  codePT?: string = null;
  prodiName?: string = null;
  prodiUrl?: string = null;
  sintaScore3Yr?: string = null;
  sintaScoreOverall?: string = null;
  subject?: object[] = [];
  scopusHIndex?: string = null;
  wosHIndex?: string = null;
  googleScholarHIndex?: string = null;
}